+++ cpp/src/io/copc_writer_public.cpp
@@ -72,7 +72,7 @@ void Writer::Close()
 bool Writer::PageExists(const VoxelKey &key) { return hierarchy_->PageExists(key); }
 
 // Writes a node to the file and reference it in the hierarchy and in the parent page
-Node Writer::DoAddNode(const VoxelKey &key, std::vector<char> in, uint64_t point_count, bool compressed_data,
+Node Writer::DoAddNode(const VoxelKey &key, const std::vector<char> &in, int32_t point_count, bool compressed_data,
                        const VoxelKey &page_key)
 {
     if (!page_key.IsValid() || !key.IsValid())
@@ -82,7 +82,7 @@ Node Writer::DoAddNode(const VoxelKey &key, std::vector<char> in, uint64_t point
     if (!key.ChildOf(page_key))
         throw std::runtime_error("Target key " + key.ToString() + " is not a child of page node " + key.ToString());
 
-    Entry e = writer_->WriteNode(std::move(in), point_count, compressed_data);
+    Entry e = writer_->WriteNode(in, point_count, compressed_data);
     e.key = key;
 
     auto node = std::make_shared<Node>(e, page_key);
@@ -123,7 +123,7 @@ Node Writer::AddNode(const VoxelKey &key, std::vector<char> const &uncompressed_
     return DoAddNode(key, uncompressed_data, 0, false, page_key);
 }
 
-Node Writer::AddNodeCompressed(const VoxelKey &key, std::vector<char> const &compressed_data, uint64_t point_count,
+Node Writer::AddNodeCompressed(const VoxelKey &key, std::vector<char> const &compressed_data, int32_t point_count,
                                const VoxelKey &page_key)
 {
     if (point_count == 0)