+++ cpp/src/las/laz_config.cpp
@@ -1,14 +1,15 @@
 #include "copc-lib/las/laz_config.hpp"
 
 #include <stdexcept>
+#include <utility>
 
 #include "copc-lib/copc/copc_config.hpp"
 
 namespace copc::las
 {
-LazConfig::LazConfig(const int8_t &point_format_id, const Vector3 &scale, const Vector3 &offset, const std::string &wkt,
+LazConfig::LazConfig(const int8_t &point_format_id, const Vector3 &scale, const Vector3 &offset, std::string wkt,
                      const las::EbVlr &extra_bytes_vlr, bool copc_flag)
-    : wkt_(wkt)
+    : wkt_(std::move(wkt))
 {
     if (point_format_id < 6 || point_format_id > 8)
         throw std::runtime_error("LasConfig: Supported point formats are 6 to 8.");