+++ example/example-laz-writer.cpp
@@ -3,7 +3,6 @@
 
 #include <copc-lib/geometry/vector3.hpp>
 #include <copc-lib/io/laz_writer.hpp>
-//#include <copc-lib/las/laz_config.hpp>
 
 using namespace copc;
 using namespace std;
@@ -22,12 +21,9 @@ las::Points RandomPoints(const las::LasHeader &header, int number_points)
 {
 
     // Random num generators between the min and max spatial bounds of the las header
-    std::uniform_int_distribution<> rand_x(header.ApplyInverseScaleX(header.min.x),
-                                           header.ApplyInverseScaleX(header.max.x));
-    std::uniform_int_distribution<> rand_y(header.ApplyInverseScaleY(header.min.y),
-                                           header.ApplyInverseScaleY(header.max.y));
-    std::uniform_int_distribution<> rand_z(header.ApplyInverseScaleZ(header.min.z),
-                                           header.ApplyInverseScaleZ(header.max.z));
+    std::uniform_int_distribution<> rand_x(header.RemoveScaleX(header.min.x), header.RemoveScaleX(header.max.x));
+    std::uniform_int_distribution<> rand_y(header.RemoveScaleY(header.min.y), header.RemoveScaleY(header.max.y));
+    std::uniform_int_distribution<> rand_z(header.RemoveScaleZ(header.min.z), header.RemoveScaleZ(header.max.z));
 
     // Create a Points object based on the LAS header
     las::Points points(header);