+++ test/reader_node_test.cpp
@@ -58,7 +58,7 @@ TEST_CASE("GetPointDataCompressed Test", "[Reader] ")
         auto hier_entry = reader.FindNode(key);
 
         auto point_vec = reader.GetPointDataCompressed(hier_entry);
-        REQUIRE(point_vec.size() > 0);
+        REQUIRE(!point_vec.empty());
     }
 }